+++ lib/common/common/src/lib.rs
@@ -11,8 +11,6 @@ pub mod cpu_utilization;
 pub mod defaults;
 pub mod delta_pack;
 pub mod disk;
-#[cfg(not(target_os = "windows"))]
-pub mod disk_cache;
 pub mod either_variant;
 pub mod ext;
 pub mod fixed_length_priority_queue;
